Restructure files we have changed the file structure according to sugestion to make everything a little bit clearer. We have also deleted an unnecessary module, and added a little comment for clarification to the top of out processWords.js file relates to #71 #63
