+++ tests/front-end-tests/test.html
@@ -10,7 +10,7 @@
   <div id="qunit"></div>
   <div id="qunit-fixture"></div>
   <script src="https://code.jquery.com/qunit/qunit-2.1.1.js"></script>
-  <script src="../public/index.js" data-cover></script>
+  <script src="../../public/index.js" data-cover></script>
   <script src="test.js"></script>
 </body>
 </html>
